start the smallest breakpoint at 0 and use the upper bound to set the canvas size
