+++ apps/designer/app/designer/features/breakpoints/use-subscribe-breakpoints.ts
@@ -3,6 +3,7 @@ import { useSubscribe } from "~/shared/pubsub";
 import { useSelectedBreakpoint } from "~/designer/shared/nano-states";
 import { useBreakpoints } from "~/shared/nano-states";
 import { utils } from "@webstudio-is/project";
+
 export const useSubscribeBreakpoints = () => {
   const [breakpoints, setBreakpoints] = useBreakpoints();
   const [selectedBreakpoint, setSelectedBreakpoint] = useSelectedBreakpoint();
