+++ packages/react-sdk/src/css/breakpoints.ts
@@ -3,7 +3,7 @@ import type { Breakpoint } from "@webstudio-is/css-data";
 export type BaseBreakpoint = Pick<Breakpoint, "label" | "minWidth">;
 
 export const initialBreakpoints: Array<BaseBreakpoint> = [
-  { label: "Mobile", minWidth: 360 },
+  { label: "Mobile", minWidth: 0 },
   { label: "Tablet", minWidth: 768 },
   { label: "Laptop", minWidth: 1024 },
   { label: "Desktop", minWidth: 1280 },
